anomaly-worker/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
python-dotenv==1.0.1
